+++ packages/typescript/index.js
@@ -4,8 +4,7 @@
  * @see https://github.com/s-panferov/awesome-typescript-loader
  */
 
-const CheckerPlugin = require('awesome-typescript-loader').CheckerPlugin
-const TsConfigPathsPlugin = require('awesome-typescript-loader').TsConfigPathsPlugin
+const { CheckerPlugin, TsConfigPathsPlugin } = require('awesome-typescript-loader')
 
 module.exports = typescript
 
@@ -20,7 +19,7 @@ function typescript (options = {}) {
     },
     module: {
       rules: [
-        {
+        Object.assign({
           test: context.fileType('application/x-typescript'),
           use: [
             {
@@ -28,7 +27,7 @@ function typescript (options = {}) {
               options
             }
           ]
-        }
+        }, context.match)
       ]
     },
     plugins: [